Add simple link to bring user to Google OAuth flow
- At this point, the user is able to select a Gmail email address to
authenticate themselves in the OAuth flow.

- Upon clicking their Gmail email address, they will be brought to
a screen showing text of: "Cannot GET /auth/google/callback". This is
expected at this point, and is ok for now. This also indicates that there
were no errors with authentication, which means that the user is
successfully "logged in" our app.

- Navigate to localhost:3000/api/current_user and you can see
user credentials. This is a further indication that the user is
successfully "logged in".

- To log the user out of our app, navigate to:
localhost:3000/api/logout

- Check if the user is successfully logged out by navigating to
localhost:3000/api/current_user again. If the screen is blank, then
the user has been logged out of our app.